Rename localstorage file to storage
